feat(era): Remove `start_from` from `EraClient` and use it instead of last file index
Improves robustness as:
* There is a single source of truth for the start index
* Counts missing files as a difference of last and current index
  to prevent freezes due to dirty download directory on restart
* Changes index type `u64` => `usize` to reduce the number of
  conversions

+++ crates/era-downloader/src/lib.rs
@@ -42,4 +42,4 @@ pub use client::{EraClient, HttpClient};
 pub use fs::read_dir;
 pub use stream::{EraMeta, EraStream, EraStreamConfig};
 
-pub(crate) const BLOCKS_PER_FILE: u64 = 8192;
+pub(crate) const BLOCKS_PER_FILE: usize = 8192;
